+++ app/src/main/java/org/oxycblt/auxio/detail/AlbumDetailFragment.kt
@@ -38,12 +38,14 @@ import org.oxycblt.auxio.music.Music
 import org.oxycblt.auxio.music.MusicParent
 import org.oxycblt.auxio.music.Song
 import org.oxycblt.auxio.playback.state.PlaybackMode
+import org.oxycblt.auxio.settings.Settings
 import org.oxycblt.auxio.ui.Header
 import org.oxycblt.auxio.ui.Item
 import org.oxycblt.auxio.ui.MenuFragment
 import org.oxycblt.auxio.util.applySpans
 import org.oxycblt.auxio.util.canScroll
 import org.oxycblt.auxio.util.collectWith
+import org.oxycblt.auxio.util.context
 import org.oxycblt.auxio.util.launch
 import org.oxycblt.auxio.util.logD
 import org.oxycblt.auxio.util.logEOrThrow
@@ -62,6 +64,7 @@ class AlbumDetailFragment :
 
     private val args: AlbumDetailFragmentArgs by navArgs()
     private val detailAdapter = AlbumDetailAdapter(this)
+    private val settings: Settings by lifecycleObject { binding -> Settings(binding.context) }
 
     override fun onCreateBinding(inflater: LayoutInflater) = FragmentDetailBinding.inflate(inflater)
 
@@ -118,7 +121,7 @@ class AlbumDetailFragment :
 
     override fun onItemClick(item: Item) {
         if (item is Song) {
-            playbackModel.play(item, PlaybackMode.IN_ALBUM)
+            playbackModel.play(item, settings.detailPlaybackMode ?: PlaybackMode.IN_ALBUM)
         }
     }