+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### What's New
 - Added a shuffle shortcut
+- You can now customize what occurs when a song is played from an album/artist/genre [#164]
 
 #### What's Fixed
 - Fixed broken tablet layouts